requirements.txt
python-docx>=1.1.0
Pillow>=10.0.0
Flask>=3.0.0
streamlit>=1.30.0
